finish filter benchmarks with both source and destination filter.  Split out EchoMessage class into separate Header file since it is used by both filterBenchmarks and EchoMessage Benchmarks.
