+++ benchmarks/helics/echoMessageBenchmarks.cpp
@@ -5,9 +5,7 @@ the top-level NOTICE for additional details. All rights reserved.
 SPDX-License-Identifier: BSD-3-Clause
 */
 
-#include "helics/application_api/Endpoints.hpp"
-#include "helics/application_api/MessageFederate.hpp"
-#include "helics/core/ActionMessage.hpp"
+#include "EchoMessage.hpp"
 #include "helics/core/BrokerFactory.hpp"
 #include "helics/core/CoreFactory.hpp"
 #include <benchmark/benchmark.h>
@@ -20,151 +18,8 @@ SPDX-License-Identifier: BSD-3-Clause
 
 #include "helics/helics-config.h"
 
-using helics::operator"" _t ;
 // static constexpr helics::Time tend = 3600.0_t;  // simulation end time
 using namespace helics;
-/** class implementing the hub for an echo test*/
-class EchoMessageHub
-{
-  public:
-    helics::Time finalTime = helics::Time (100, time_units::ms);  // final time
-  private:
-    std::unique_ptr<helics::MessageFederate> mFed;
-    helics::Endpoint ept;
-    bool initialized = false;
-    bool readyToRun = false;
-
-  public:
-    EchoMessageHub () = default;
-
-    void run (std::function<void ()> callOnReady = {})
-    {
-        if (!readyToRun)
-        {
-            makeReady ();
-        }
-        if (callOnReady)
-        {
-            callOnReady ();
-        }
-        mainLoop ();
-    };
-
-    void initialize (const std::string &coreName)
-    {
-        std::string name = "echohub";
-        helics::FederateInfo fi;
-        fi.coreName = coreName;
-        mFed = std::make_unique<helics::MessageFederate> (name, fi);
-        ept = mFed->registerGlobalEndpoint ("echo");
-        initialized = true;
-    }
-
-    void makeReady ()
-    {
-        if (!initialized)
-        {
-            throw ("must initialize first");
-        }
-        mFed->enterExecutingMode ();
-        readyToRun = true;
-    }
-
-    void mainLoop ()
-    {
-        auto cTime = 0.0_t;
-        while (cTime <= finalTime)
-        {
-            while (ept.hasMessage ())
-            {
-                auto m = ept.getMessage ();
-                std::swap (m->source, m->dest);
-                std::swap (m->original_source, m->original_dest);
-                ept.send (std::move (m));
-            }
-
-            cTime = mFed->requestTime (finalTime + 0.05);
-        }
-        mFed->finalize ();
-    }
-};
-
-class EchoMessageLeaf
-{
-  private:
-    std::unique_ptr<helics::MessageFederate> mFed;
-    helics::Endpoint ept;
-
-    int index_ = 0;
-    bool initialized = false;
-    bool readyToRun = false;
-
-  public:
-    EchoMessageLeaf () = default;
-
-    void run (std::function<void ()> callOnReady = {})
-    {
-        if (!readyToRun)
-        {
-            makeReady ();
-        }
-        if (callOnReady)
-        {
-            callOnReady ();
-        }
-        mainLoop ();
-    };
-    void initialize (const std::string &coreName, int index)
-    {
-        std::string name = "echoleaf_" + std::to_string (index);
-        index_ = index;
-        helics::FederateInfo fi;
-        fi.coreName = coreName;
-        mFed = std::make_unique<helics::MessageFederate> (name, fi);
-        // this is a local endpoint
-        ept = mFed->registerEndpoint ("leaf");
-        initialized = true;
-    }
-
-    void makeReady ()
-    {
-        if (!initialized)
-        {
-            throw ("must initialize first");
-        }
-        mFed->enterExecutingMode ();
-        readyToRun = true;
-    }
-
-    void mainLoop ()
-    {
-        int cnt = 0;
-        // this is  to make a fixed size string that is different for each federate but has sufficient length to
-        // get beyond SSO
-        const std::string txstring = std::to_string (100000 + index_) + std::string (100, '1');
-        const int iter = 5000;
-        while (cnt <= iter + 1)
-        {
-            mFed->requestNextStep ();
-            ++cnt;
-            if (cnt <= iter)
-            {
-                ept.send ("echo", txstring);
-            }
-            while (ept.hasMessage ())
-            {
-                auto m = ept.getMessage ();
-                auto &nstring = m->data.to_string ();
-                if (nstring != txstring)
-                {
-                    std::cout << "incorrect string\n";
-                    break;
-                }
-            }
-        }
-        mFed->finalize ();
-    }
-};
 
 static void BM_echo_singleCore (benchmark::State &state)
 {
@@ -173,7 +28,7 @@ static void BM_echo_singleCore (benchmark::State &state)
         state.PauseTiming ();
 
         int feds = static_cast<int> (state.range (0));
-        gmlc::concurrency::Barrier brr (feds + 1);
+        gmlc::concurrency::Barrier brr (static_cast<size_t>(feds) + 1);
         auto wcore = helics::CoreFactory::create (core_type::INPROC, std::string ("--autobroker --federates=") +
                                                                        std::to_string (feds + 1));
         EchoMessageHub hub;
@@ -219,7 +74,7 @@ static void BM_echo_multiCore (benchmark::State &state, core_type cType)
         state.PauseTiming ();
 
         int feds = static_cast<int> (state.range (0));
-        gmlc::concurrency::Barrier brr (feds + 1);
+        gmlc::concurrency::Barrier brr (static_cast<size_t>(feds) + 1);
 
         auto broker = helics::BrokerFactory::create (cType, "brokerb",
                                                      std::string ("--federates=") + std::to_string (feds + 1));
